# openapi/planning-activities.yaml
openapi: 3.1.0
info:
  title: Planning Activities API
  version: 0.1.0
  description: |
    REST-Schnittstelle für die Gantt-Aktivitäten der Order-Management-Anwendung.
    Alle Clients arbeiten gegen denselben Datenbestand. Validierungen für Orts-,
    Kapazitäts-, Arbeitszeit- und Qualifikationskonflikte werden serverseitig
    ausgeführt und können gezielt angefordert werden.
servers:
  - url: /api/v1
    description: Default Backend-URL (überschreibbar per API_CONFIG).
paths:
...
  /api/templates/{templateId}/timeline:
    get:
      summary: Template-Timeline laden (Stage + LOD)
      description: >
        Liefert Activities oder Service-Aggregationen für eine Template-Timeline (eigene Tabelle je Template).
        Query-Parameter `stage=base|operations`, `lod=activity|service`, `from`, `to` (ISO).
      parameters:
        - in: path
          name: templateId
          required: true
          schema:
            type: string
        - in: query
          name: from
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: stage
          required: false
          schema:
            type: string
            enum: [base, operations]
          description: Stage-Kontext innerhalb des Templates (default base).
        - in: query
          name: lod
          required: false
          schema:
            type: string
            enum: [activity, service]
          description: Level of Detail.
      responses:
        '200':
          description: Template-Timeline-Daten
          content:
            application/json:
              schema:
                $ref: './planning/components/schemas.yaml#/schemas/TimelineResponse'
  /api/templates:
    get:
      summary: Templates auflisten
      description: Liefert Metadaten aller Templates (eigene Timeline-Tabellen).
      responses:
        '200':
          description: Liste der Templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: './planning/components/schemas.yaml#/schemas/TemplateSet'
    post:
      summary: Template anlegen
      description: Legt ein neues Template an und erzeugt die zugehörige Timeline-Tabelle.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './planning/components/schemas.yaml#/schemas/CreateTemplateSetRequest'
      responses:
        '201':
          description: Angelegtes Template
          content:
            application/json:
              schema:
                $ref: './planning/components/schemas.yaml#/schemas/TemplateSet'
  /api/templates/{templateId}:
    get:
      summary: Template laden
      parameters:
        - in: path
          name: templateId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Template-Metadaten
          content:
            application/json:
              schema:
                $ref: './planning/components/schemas.yaml#/schemas/TemplateSet'
    put:
      summary: Template aktualisieren
      parameters:
        - in: path
          name: templateId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './planning/components/schemas.yaml#/schemas/UpdateTemplateSetRequest'
      responses:
        '200':
          description: Aktualisiertes Template
          content:
            application/json:
              schema:
                $ref: './planning/components/schemas.yaml#/schemas/TemplateSet'
    delete:
      summary: Template löschen
      parameters:
        - in: path
          name: templateId
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Gelöscht, zugehörige Timeline-Tabelle entfernt
  /api/templates/{templateId}/activities/{activityId}:
    put:
      summary: Template-Aktivität anlegen/aktualisieren
      parameters:
        - in: path
          name: templateId
          required: true
          schema:
            type: string
        - in: path
          name: activityId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
            application/json:
              schema:
                $ref: './planning/components/schemas.yaml#/schemas/TemplateActivityUpsert'
      responses:
        '200':
          description: Gespeicherte Aktivität
          content:
            application/json:
              schema:
                $ref: './planning/components/schemas.yaml#/schemas/TimelineActivity'
    delete:
      summary: Template-Aktivität löschen
      parameters:
        - in: path
          name: templateId
          required: true
          schema:
            type: string
        - in: path
          name: activityId
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Aktivität entfernt
  /api/templates/{templateId}/rollout:
    post:
      summary: Template in Live-Planning ausrollen
      description: Kopiert die Template-Aktivitäten in die Live-`activities`-Tabelle (Stage wählbar, optionaler Zeitanker).
      parameters:
        - in: path
          name: templateId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stage:
                  type: string
                  enum: [base, operations]
                anchorStart:
                  type: string
                  format: date-time
                  description: Optionaler Ankerzeitpunkt, relativ werden alle Aktivitäten verschoben.
              required: [stage]
      responses:
        '200':
          description: Ausgerollte Aktivitäten (Live)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: './planning/components/schemas.yaml#/schemas/TimelineActivity'

# WebSocket-Notizen (AsyncAPI empfohlen, hier nur informativ):
# - Live-Timeline WS: Namespace /timeline
#   Eingehend: VIEWPORT_CHANGED {from,to,lod,stage}, ACTIVITY_UPDATE_REQUEST {activityId,newStart,newEnd,stage}
#   Ausgehend: ACTIVITY_UPDATE_ACCEPTED, ACTIVITY_UPDATED, SERVICE_UPDATED, ABSENCE_UPDATED, ACTIVITY_UPDATE_VALIDATION_RESULT
# - Template-Timeline WS: Namespace /templates
#   Kontext zusätzlich templateId; gleiche Message-Typen wie Live-Timeline.

# WebSocket-Hinweise (nicht formal als OpenAPI-WebSocket beschrieben):
# - Live-Timeline WS: Namespace /timeline, Messages: VIEWPORT_CHANGED (stage,lod,from,to), ACTIVITY_UPDATE_REQUEST, ACTIVITY_UPDATED/SERVICE_UPDATED Broadcasts.
# - Template-Timeline WS: Namespace /templates, Kontext zusätzlich templateId; gleiche Message-Typen wie Live.
...